refactor(gui): split code by category
