+++ include/LCUI/gui/Makefile.am
@@ -4,5 +4,7 @@ SUBDIRS=widget
 pkginclude_HEADERS = widget_base.h widget_task.h widget_prototype.h \
 widget_style.h widget_event.h widget_paint.h widget.h css_library.h \
 widget_helper.h css_parser.h css_rule_font_face.h css_fontstyle.h \
-builder.h metrics.h widget_layout.h
+builder.h metrics.h widget_layout.h widget_attribute.h widget_id.h \
+widget_class.h widget_status.h widget_tree.h
+
 pkgincludedir=$(prefix)/include/LCUI/gui
