+++ include/LCUI/gui/widget_prototype.h
@@ -42,6 +42,13 @@ LCUI_API LCUI_WidgetPrototype LCUIWidget_GetPrototype(const char *name);
 LCUI_API LCUI_WidgetPrototype LCUIWidget_NewPrototype(const char *name,
 						      const char *parent_name);
 
+/** 判断部件类型 */
+LCUI_API LCUI_BOOL Widget_CheckType(LCUI_Widget w, const char *type);
+
+/** 判断部件原型 */
+LCUI_API LCUI_BOOL Widget_CheckPrototype(LCUI_Widget w,
+					 LCUI_WidgetPrototypeC proto);
+
 LCUI_API void *Widget_GetData(LCUI_Widget widget, LCUI_WidgetPrototype proto);
 
 LCUI_API void *Widget_AddData(LCUI_Widget widget,
